refactor: mconsoldiate several common transforms into their own  file to reduce repitiveness
